+++ TargetedSpellsBarMixin.lua
@@ -7,6 +7,7 @@ TargetedSpellsBarMixin = {}
 function TargetedSpellsBarMixin:OnLoad()
 	self.Bar:SetStatusBarTexture("")
 	self.ProgressBar:SetStatusBarTexture("Interface\\RaidFrame\\Raid-Bar-Hp-Fill")
+	self.elapsed = 0
 	Private.EventRegistry:RegisterCallback(Private.Enum.Events.SETTING_CHANGED, self.OnSettingChanged, self)
 end
 
@@ -25,35 +26,40 @@ function TargetedSpellsBarMixin:OnSizeChanged()
 	self.Icon:ClearAllPoints()
 	self.ProgressBar:ClearAllPoints()
 	self.ProgressBar.SpellName:ClearAllPoints()
+	self.ProgressBar.Divider:ClearAllPoints()
 	self.ProgressBar.TargetName:ClearAllPoints()
 
-	-- The leftmost slot frame that ProgressBar and SpellName anchor their left edge to
 	local slotFrame = nil
 
 	if showTargetMarker and showIcon then
-		self.CustomElementsFrame.TargetMarker:SetPoint("TOPLEFT", self, "TOPLEFT", 0, 0)
-		self.CustomElementsFrame.TargetMarker:SetPoint("BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
-		self.Icon:SetPoint("TOPLEFT", self.CustomElementsFrame.TargetMarker, "TOPRIGHT", 0, 0)
-		self.Icon:SetPoint("BOTTOMLEFT", self.CustomElementsFrame.TargetMarker, "BOTTOMRIGHT", 0, 0)
+		PixelUtil.SetPoint(self.CustomElementsFrame.TargetMarker, "TOPLEFT", self, "TOPLEFT", 0, 0)
+		PixelUtil.SetPoint(self.CustomElementsFrame.TargetMarker, "BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
+		PixelUtil.SetPoint(self.Icon, "TOPLEFT", self.CustomElementsFrame.TargetMarker, "TOPRIGHT", 0, 0)
+		PixelUtil.SetPoint(self.Icon, "BOTTOMLEFT", self.CustomElementsFrame.TargetMarker, "BOTTOMRIGHT", 0, 0)
+
 		slotFrame = self.Icon
 	elseif showTargetMarker then
-		self.CustomElementsFrame.TargetMarker:SetPoint("TOPLEFT", self, "TOPLEFT", 0, 0)
-		self.CustomElementsFrame.TargetMarker:SetPoint("BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
+		PixelUtil.SetPoint(self.CustomElementsFrame.TargetMarker, "TOPLEFT", self, "TOPLEFT", 0, 0)
+		PixelUtil.SetPoint(self.CustomElementsFrame.TargetMarker, "BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
+
 		slotFrame = self.CustomElementsFrame.TargetMarker
 	elseif showIcon then
-		self.Icon:SetPoint("TOPLEFT", self, "TOPLEFT", 0, 0)
-		self.Icon:SetPoint("BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
+		PixelUtil.SetPoint(self.Icon, "TOPLEFT", self, "TOPLEFT", 0, 0)
+		PixelUtil.SetPoint(self.Icon, "BOTTOMLEFT", self, "BOTTOMLEFT", 0, 0)
+
 		slotFrame = self.Icon
 	end
 
 	local showSpellName = TargetedSpellsSaved.Settings.Party.FeatureFlags[Private.Enum.FeatureFlag.ShowSpellName]
+	local hasDivider = TargetedSpellsSaved.Settings.Party.NameDivider ~= Private.Enum.NameDivider.None
 
 	if slotFrame then
-		self.ProgressBar:SetPoint("TOPLEFT", slotFrame, "TOPRIGHT", 0, 0)
-		self.ProgressBar:SetPoint("BOTTOMRIGHT", self, "BOTTOMRIGHT", 0, 0)
-		self.ProgressBar.SpellName:SetPoint("LEFT", slotFrame, "RIGHT", 4, 0)
+		PixelUtil.SetPoint(self.ProgressBar, "TOPLEFT", slotFrame, "TOPRIGHT", 0, 0)
+		PixelUtil.SetPoint(self.ProgressBar, "BOTTOMRIGHT", self, "BOTTOMRIGHT", 0, 0)
+		PixelUtil.SetPoint(self.ProgressBar.SpellName, "LEFT", slotFrame, "RIGHT", 4, 0)
+
 		if not showSpellName then
-			self.ProgressBar.TargetName:SetPoint("LEFT", slotFrame, "RIGHT", 4, 0)
+			PixelUtil.SetPoint(self.ProgressBar.TargetName, "LEFT", slotFrame, "RIGHT", 4, 0)
 		end
 	else
 		self.ProgressBar:SetAllPoints(self)
@@ -64,11 +70,18 @@ function TargetedSpellsBarMixin:OnSizeChanged()
 	end
 
 	if showSpellName then
-		self.ProgressBar.TargetName:SetPoint("LEFT", self.ProgressBar.SpellName, "RIGHT", 0, 0)
+		if hasDivider then
+			PixelUtil.SetPoint(self.ProgressBar.Divider, "LEFT", self.ProgressBar.SpellName, "RIGHT", 0, 0)
+			PixelUtil.SetPoint(self.ProgressBar.TargetName, "LEFT", self.ProgressBar.Divider, "RIGHT", 0, 0)
+		else
+			PixelUtil.SetPoint(self.ProgressBar.TargetName, "LEFT", self.ProgressBar.SpellName, "RIGHT", 0, 0)
+		end
 	end
 
 	self.ProgressBar.SpellName:SetWidth(TargetedSpellsSaved.Settings.Party.SpellNameWidth)
 	self.ProgressBar.TargetName:SetWidth(TargetedSpellsSaved.Settings.Party.TargetNameWidth)
+
+	self:SetDivider()
 end
 
 function TargetedSpellsBarMixin:OnSettingChanged(key, flagIdOrValue, newBool)
@@ -76,22 +89,43 @@ function TargetedSpellsBarMixin:OnSettingChanged(key, flagIdOrValue, newBool)
 		PixelUtil.SetSize(self, flagIdOrValue, TargetedSpellsSaved.Settings.Party.Height)
 	elseif key == Private.Settings.Keys.Party.Height then
 		PixelUtil.SetSize(self, TargetedSpellsSaved.Settings.Party.Width, flagIdOrValue)
+	elseif key == Private.Settings.Keys.Party.ShowDuration then
+		self:SetShowDuration(flagIdOrValue)
 	elseif key == Private.Settings.Keys.Party.FeatureFlags then
 		if
 			flagIdOrValue == Private.Enum.FeatureFlag.ShowIcon
 			or flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetMarker
 			or flagIdOrValue == Private.Enum.FeatureFlag.ShowSpellName
 			or flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetName
+			or flagIdOrValue == Private.Enum.FeatureFlag.ShowDuration
 		then
-			if flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetMarker and not newBool then
-				self.CustomElementsFrame.TargetMarker:Hide()
+			if flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetMarker then
+				self.CustomElementsFrame.TargetMarker:SetShown(newBool)
+				self:OnSizeChanged()
+			elseif flagIdOrValue == Private.Enum.FeatureFlag.ShowIcon then
+				self.Icon:SetShown(newBool)
+				self:OnSizeChanged()
+			elseif flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetName then
+				self.ProgressBar.TargetName:SetShown(newBool)
+				self:SetDivider()
+				self:OnSizeChanged()
+			elseif flagIdOrValue == Private.Enum.FeatureFlag.ShowSpellName then
+				self.ProgressBar.SpellName:SetShown(newBool)
+				self:SetDivider()
+				self:OnSizeChanged()
+			elseif flagIdOrValue == Private.Enum.FeatureFlag.ShowDuration then
+				self:SetShowDuration(newBool)
 			end
-			if flagIdOrValue == Private.Enum.FeatureFlag.ShowTargetName and not newBool then
-				self.ProgressBar.TargetName:Hide()
-			end
-
-			self:OnSizeChanged()
 		end
+	elseif
+		key == Private.Settings.Keys.Party.FontSize
+		or key == Private.Settings.Keys.Party.Font
+		or key == Private.Settings.Keys.Party.FontFlags
+	then
+		self:SetFont()
+	elseif key == Private.Settings.Keys.Party.NameDivider then
+		self:SetDivider()
+		self:OnSizeChanged()
 	elseif key == Private.Settings.Keys.Party.SpellNameWidth then
 		self.ProgressBar.SpellName:SetWidth(flagIdOrValue)
 	elseif key == Private.Settings.Keys.Party.TargetNameWidth then
@@ -99,7 +133,40 @@ function TargetedSpellsBarMixin:OnSettingChanged(key, flagIdOrValue, newBool)
 	end
 end
 
-function TargetedSpellsBarMixin:OnUpdate() end
+function TargetedSpellsBarMixin:SetDivider()
+	if
+		not TargetedSpellsSaved.Settings.Party.FeatureFlags[Private.Enum.FeatureFlag.ShowSpellName]
+		or not TargetedSpellsSaved.Settings.Party.FeatureFlags[Private.Enum.FeatureFlag.ShowTargetName]
+	then
+		self.ProgressBar.Divider:Hide()
+		return
+	end
+
+	local nameDivider = TargetedSpellsSaved.Settings.Party.NameDivider
+
+	if nameDivider == Private.Enum.NameDivider.None then
+		self.ProgressBar.Divider:Hide()
+		return
+	end
+
+	if
+		nameDivider == Private.Enum.NameDivider.Arrow
+		or nameDivider == Private.Enum.NameDivider.Arrows
+		or nameDivider == Private.Enum.NameDivider.Pipe
+	then
+		self.ProgressBar.Divider:SetText(" " .. nameDivider .. " ")
+	elseif nameDivider == Private.Enum.NameDivider.Colon then
+		self.ProgressBar.Divider:SetText(nameDivider .. " ")
+	else
+		self.ProgressBar.Divider:SetText(nameDivider)
+	end
+
+	self.ProgressBar.Divider:Show()
+end
+
+function TargetedSpellsBarMixin:SetShowDuration(showDuration)
+	self.ProgressBar.Duration:SetShown(showDuration)
+end
 
 function TargetedSpellsBarMixin:Reset()
 	self:SetParent(UIParent)
@@ -108,6 +175,7 @@ function TargetedSpellsBarMixin:Reset()
 	self:ClearAllPoints()
 	self.startTime = nil
 	self.CustomElementsFrame.TargetMarker:Hide()
+	self.ProgressBar.Divider:Hide()
 	self.ProgressBar.TargetName:Hide()
 	self:Hide()
 end
@@ -155,11 +223,7 @@ do
 			color = whiteDefaultColor
 		end
 
-		self.ProgressBar.TargetName:SetFormattedText(
-			"%s%s",
-			TargetedSpellsSaved.Settings.Party.FeatureFlags[Private.Enum.FeatureFlag.ShowSpellName] and " >> " or "",
-			name
-		)
+		self.ProgressBar.TargetName:SetText(name)
 		self.ProgressBar.TargetName:SetTextColor(color.r, color.g, color.b, color.a)
 		self.ProgressBar.TargetName:Show()
 	end
@@ -203,12 +267,13 @@ function TargetedSpellsBarMixin:SetSpellId(spellId)
 	self.Icon:SetTexture(texture)
 
 	if TargetedSpellsSaved.Settings.Party.FeatureFlags[Private.Enum.FeatureFlag.ShowSpellName] then
-		local name = spellId == nil and addonName or C_Spell.GetSpellName(spellId)
-		self.ProgressBar.SpellName:SetText(name)
+		self.ProgressBar.SpellName:SetText(spellId == nil and addonName or C_Spell.GetSpellName(spellId))
 		self.ProgressBar.SpellName:Show()
 	else
 		self.ProgressBar.SpellName:Hide()
 	end
+
+	self:SetDivider()
 
 	if spellId == nil then
 		return
@@ -298,6 +363,88 @@ function TargetedSpellsBarMixin:IsSpellImportant(boolOverride)
 end
 
 function TargetedSpellsBarMixin:SetDuration(duration)
-	self.duration = duration
+	self.ProgressBar.Duration:SetText("")
 	self.ProgressBar:SetTimerDuration(duration)
+end
+
+function TargetedSpellsBarMixin:SetFont()
+	local hasShadow = TargetedSpellsSaved.Settings.Party.FontFlags[Private.Enum.FontFlags.SHADOW]
+
+	for _, fontString in ipairs({
+		self.ProgressBar.SpellName,
+		self.ProgressBar.Divider,
+		self.ProgressBar.TargetName,
+		self.ProgressBar.Duration,
+	}) do
+		fontString:SetFont(
+			TargetedSpellsSaved.Settings.Party.Font,
+			TargetedSpellsSaved.Settings.Party.FontSize,
+			TargetedSpellsSaved.Settings.Party.FontFlags[Private.Enum.FontFlags.OUTLINE] and "OUTLINE" or ""
+		)
+
+		fontString:SetShadowOffset(hasShadow and 1 or 0, hasShadow and -1 or 0)
+
+		if hasShadow then
+			fontString:SetShadowColor(0, 0, 0, 1)
+		end
+	end
+end
+
+do
+	local formatter = C_StringUtil.CreateNumericRuleFormatter()
+
+	formatter:SetBreakpoints({
+		{
+			threshold = 0,
+			rounding = Enum.NumericRuleFormatRounding.Nearest,
+			format = "%.1f",
+			step = 0.1,
+		},
+		{
+			threshold = 3,
+			rounding = Enum.NumericRuleFormatRounding.Nearest,
+			format = "%d",
+		},
+		{
+			threshold = 60,
+			rounding = Enum.NumericRuleFormatRounding.Nearest,
+			format = "%d:%02d",
+			components = {
+				{
+					div = 60,
+				},
+				{
+					mod = 60,
+				},
+			},
+		},
+		{
+			threshold = 300,
+			rounding = Enum.NumericRuleFormatRounding.Up,
+			format = "%dm",
+			components = {
+				{
+					div = 60,
+				},
+			},
+		},
+	})
+
+	function TargetedSpellsBarMixin:OnUpdate(elapsed)
+		self.elapsed = self.elapsed + elapsed
+
+		if self.elapsed < 0.1 then
+			return
+		end
+
+		self.elapsed = self.elapsed - 0.1
+
+		local duration = self.ProgressBar:GetTimerDuration()
+
+		if duration == nil then
+			return
+		end
+
+		self.ProgressBar.Duration:SetText(duration:FormatRemainingDuration(formatter))
+	end
 end